Move compiler tests.
